Reverted the interface renaming done in #50
All code in the bundle is still relying on the transformMarkdown method in
parsers, not on the transform method.
Normal parsers are still working because they still have the method, and
PHP does not have strict type checking for typehints.
However, the SundownParser was still broken. #51 renamed the method to
avoid a fatal error for the interface, but then the method used by the
bundle became missing.